refactor: move incoming state machine into lightning client crate
